requirements.txt
# Make sure to install those as additional_dependencies in the
# pre-commit hooks for pylint & mypy
# Also update the readme accordingly

# When dependencies release new versions and tests succeed, we should try to expand the allowed
# versions and only increase the lower bound if necessary

# httpx has no stable release yet, so let's be cautious for now
httpx ~= 0.23.3
python-dotenv
